+++ app/src/main/java/com/sandy/memorizingvoca/ui/feature/voca_list/navigation/VocaListNavigation.kt
@@ -22,7 +22,7 @@ fun NavController.navigateToList(
 fun NavGraphBuilder.vocaListScreen(
     onNavigateBack: () -> Unit,
     onNavigateFull: () -> Unit,
-    onNavigateQuiz1: () -> Unit,
+    onNavigateQuiz1: (Int) -> Unit,
     onNavigateQuiz2: () -> Unit,
     onVocaItemClick: (Int) -> Unit,
 ) {
